Resolve Local Storage Issue: All Data is stored in local storage, for new as well as existing user

+++ components/Routes/UserPage.jsx
@@ -1,6 +1,5 @@
 import React, { useRef, useState } from "react";
 import { useLocation } from "react-router-dom";
-import QuoteCard from "../QuoteCard";
 import Card from "react-bootstrap/Card";
 import CardGroup from "react-bootstrap/CardGroup";
 import AddQuote from "../AddQuote";
@@ -8,19 +7,36 @@ import UserQuotes from "../userQuotes";
 import QuoteCarousel from "../QuoteCarousel";
 
 function UserPage() {
-  const [userQuotes, setUserQuotes] = useState([]);
-  // const [noQuotes, setNoQuotes] = useState(true)
-  const [userInfo, setUserInfo] = useState({});
   const location = useLocation();
   const user = location.state.user;
+  const userQuotes = useRef([]);
+  const [userInfo, setUserInfo] = useState({ user, userQuotes });
+
+  // localStorage.setItem(`${user.userName}`, JSON.stringify(userInfo));
 
   const handleQuoteSubmit = (newQuotes) => {
-    setUserQuotes([...userQuotes, newQuotes]);
-    console.log(userQuotes);
-    setUserInfo({ user, userQuotes });
-    console.log(userInfo);
-    localStorage.setItem(`${user.userName}`, JSON.stringify(userInfo));
-    // setNoQuotes(false);
+    console.log(newQuotes);
+    const existingUser = JSON.parse(localStorage.getItem(`${user.userName}`));
+    // console.log(existingUser.userQuotes.current, "current");
+
+    if (existingUser) {
+      // console.log(existingUser.userQuotes.current, "inside if");
+      existingUser.userQuotes.current = [
+        ...existingUser.userQuotes.current,
+        newQuotes,
+      ];
+      console.log(existingUser.userQuotes.current, "existingUser userQuotes current");
+      userQuotes.current = existingUser.userQuotes.current;
+      console.log(userQuotes, "User Quotes");
+      setUserInfo({ user, userQuotes });
+      localStorage.setItem(`${user.userName}`, JSON.stringify(userInfo));
+      console.log(userInfo, "userInfo");
+    } else {
+      // console.log("Inside Else")
+      userQuotes.current = [...userQuotes.current, newQuotes];
+      setUserInfo({ user, userQuotes });
+      localStorage.setItem(`${user.userName}`, JSON.stringify(userInfo));
+    }
   };
   return (
     <CardGroup>
@@ -31,7 +47,7 @@ function UserPage() {
         <Card>
           <AddQuote onQuoteSubmit={handleQuoteSubmit} />
         </Card>
-      <UserQuotes user={user} />
+        <UserQuotes />
       </div>
     </CardGroup>
   );
